encode-faces/pyproject.toml
[project]
name = "encode-faces"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.22.0",
    "azure-storage-blob>=12.25.1",
    "dlib>=19.24.8",
    "face-recognition>=1.3.0",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "setuptools>=79.0.1",
]
